% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARIF.R
\name{harmonize_matrices}
\alias{harmonize_matrices}
\title{function to harmonize the matrices (Step1)}
\usage{
harmonize_matrices(matrices)
}
\arguments{
\item{matrices}{A list of two or more matrices of z-statistics of corresponding to two or more groups.}
}
\description{
function to harmonize the matrices (Step1)
}
