% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARIF.R
\name{TARIF.D}
\alias{TARIF.D}
\title{function to run JIVE (Step2)}
\usage{
TARIF.D(matrices, mi, status)
}
\arguments{
\item{mi}{The maximum number of iterations for each instance of the JIVE algorithm.}

\item{status}{A boolean indicating whether or not to give output showing the progress of the JIVE algorithm. If TRUE, the algorithm will print out updates about the number of iterations the algorithm is taking and the progress of the rank selection method, if applicable. If FALSE, the algorithms will give no printed output when run.}
}
\description{
function to run JIVE (Step2)
}
